Remove unused servlets and templates
