--- projects/src/main/java/timeboard/projects/ProjectTaskConfigServlet.java
@@ -1,380 +1,0 @@
-package timeboard.projects;
-
-/*-
- * #%L
- * projects
- * %%
- * Copyright (C) 2019 Timeboard
- * %%
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- * #L%
- */
-
-import java.io.IOException;
-import java.text.ParseException;
-import java.text.SimpleDateFormat;
-import java.time.LocalDate;
-import java.time.format.DateTimeFormatter;
-import java.util.Date;
-import java.util.LinkedHashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
-import javax.servlet.Servlet;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
-import org.osgi.service.component.annotations.ServiceScope;
-import timeboard.core.api.ProjectService;
-import timeboard.core.api.UserService;
-import timeboard.core.api.exceptions.BusinessException;
-import timeboard.core.model.*;
-import timeboard.core.ui.TimeboardServlet;
-import timeboard.core.ui.ViewModel;
-
-@Component(
-        service = Servlet.class,
-        scope = ServiceScope.PROTOTYPE,
-        property = {
-                "osgi.http.whiteboard.servlet.pattern=/projects/tasks/config",
-                "osgi.http.whiteboard.context.select=(osgi.http.whiteboard.context.name=timeboard)"
-        }
-)
-public class ProjectTaskConfigServlet extends TimeboardServlet {
-
-    private static final SimpleDateFormat DATE_FORMAT = new SimpleDateFormat("yyyy-MM-dd");
-
-    @Reference
-    public ProjectService projectService;
-
-    @Reference
-    public UserService userService;
-
-
-    @Override
-    protected ClassLoader getTemplateResolutionClassLoader() {
-        return ProjectTaskConfigServlet.class.getClassLoader();
-    }
-
-    @Override
-    protected void handleGet(User actor, HttpServletRequest request, HttpServletResponse response, ViewModel viewModel) throws ServletException, IOException, BusinessException {
-
-        Task task = new Task();
-        if (request.getParameter("taskID") != null) {
-            // Update case
-            long taskID = Long.parseLong(request.getParameter("taskID"));
-            task = (Task) this.projectService.getTaskByID(actor, taskID);
-        }
-        viewModel.getViewDatas().put("task", new TaskForm(task));
-
-        long projectID = Long.parseLong(request.getParameter("projectID"));
-        Project project = this.projectService.getProjectByID(actor, projectID);
-
-        viewModel.setTemplate("projects:details_project_tasks_config.html");
-        viewModel.getViewDatas().put("project", project);
-        viewModel.getViewDatas().put("tasks", this.projectService.listProjectTasks(actor, project));
-        viewModel.getViewDatas().put("taskTypes", this.projectService.listTaskType());
-        viewModel.getViewDatas().put("allTaskStatus", TaskStatus.values());
-        viewModel.getViewDatas().put("allProjectMilestones", this.projectService.listProjectMilestones(actor, project));
-
-
-        /* Get datas for line-chart*/
-        if (task.getId() != null) {
-            this.getDatasForCharts(actor, viewModel, task);
-        }
-
-    }
-
-    private void getDatasForCharts(User actor, ViewModel viewModel, Task task) throws BusinessException {
-
-        // Datas for dates (Axis X)
-        String formatLocalDate = "yyyy-MM-dd";
-        String formatDateToDisplay = "dd/MM/yyyy";
-        LocalDate start = LocalDate.parse(new SimpleDateFormat(formatLocalDate).format(task.getStartDate()));
-        LocalDate end = LocalDate.parse(new SimpleDateFormat(formatLocalDate).format(task.getEndDate()));
-        List<String> listOfTaskDates = start.datesUntil(end.plusDays(1))
-                .map(localDate -> localDate.format(DateTimeFormatter.ofPattern(formatDateToDisplay)))
-                .collect(Collectors.toList());
-        viewModel.getViewDatas().put("listOfTaskDates", listOfTaskDates);
-
-        // Datas for effort spent (Axis Y)
-        List<EffortHistory> effortSpentDB = this.projectService.getEffortSpentByTaskAndPeriod(actor, task, task.getStartDate(), task.getEndDate());
-        final EffortHistory[] lastEffortSpentSum = {new EffortHistory(task.getStartDate(), 0.0)};
-        Map<Date, Double> effortSpentMap = listOfTaskDates
-                .stream()
-                .map(dateString -> {
-                    try {
-                        return new SimpleDateFormat(formatDateToDisplay).parse(dateString);
-                    } catch (ParseException e) {
-                        e.printStackTrace();
-                    }
-                    return null;
-                })
-                .map(date -> effortSpentDB.stream()
-                        .filter(es -> new SimpleDateFormat(formatDateToDisplay).format(es.getDate()).equals(new SimpleDateFormat(formatDateToDisplay).format(date)))
-                        .map(effort -> {
-                            lastEffortSpentSum[0] = new EffortHistory(date, effort.getValue());
-                            return lastEffortSpentSum[0];
-                        })
-                        .findFirst().orElse(new EffortHistory(date, lastEffortSpentSum[0].getValue())))
-                .collect(Collectors.toMap(
-                        e -> e.getDate(),
-                        e -> e.getValue(),
-                        (x, y) -> y, LinkedHashMap::new
-                ));
-        viewModel.getViewDatas().put("effortSpentDatasForChart", effortSpentMap.values());
-
-        // Datas for effort estimate (Axis Y)
-        List<EffortHistory> effortLeftDB = this.projectService.getTaskEffortLeftHistory(actor, task);
-        final EffortHistory[] lastEffortEstimate = {new EffortHistory(task.getStartDate(), task.getOriginalEstimate())};
-        Map<Date, Double> effortEstimateMap = listOfTaskDates
-                .stream()
-                .map(dateString -> {
-                    try {
-                        return new SimpleDateFormat(formatDateToDisplay).parse(dateString);
-                    } catch (ParseException e) {
-                        e.printStackTrace();
-                    }
-                    return null;
-                })
-                .map(date -> effortLeftDB.stream()
-                        .filter(el -> new SimpleDateFormat(formatDateToDisplay).format(el.getDate()).equals(new SimpleDateFormat(formatDateToDisplay).format(date)))
-                        .map(effortLeft -> {
-                            lastEffortEstimate[0] = new EffortHistory(date, effortLeft.getValue() + effortSpentMap.get(date));
-                            return lastEffortEstimate[0];
-                        })
-                        .findFirst().orElse(new EffortHistory(date, lastEffortEstimate[0].getValue())))
-                .collect(Collectors.toMap(
-                        e -> e.getDate(),
-                        e -> e.getValue(),
-                        (x, y) -> y, LinkedHashMap::new
-                ));
-        viewModel.getViewDatas().put("realEffortDatasForChart", effortEstimateMap.values());
-    }
-
-    @Override
-    protected void handlePost(User actor, HttpServletRequest request, HttpServletResponse response, ViewModel viewModel) throws ServletException, IOException, BusinessException {
-
-        long projectID = Long.parseLong(request.getParameter("projectID"));
-        Project project = this.projectService.getProjectByID(actor, projectID);
-          Task currentTask = null;
-
-        try {
-
-            if (!getParameter(request, "taskID").get().isEmpty()) {
-                Long taskID = Long.parseLong(request.getParameter("taskID"));
-                currentTask = (Task) this.projectService.getTaskByID(actor, taskID);
-                currentTask = this.updateTask(actor, project, currentTask, request);
-            } else {
-                currentTask = this.createTask(actor, project, request);
-            }
-
-            viewModel.getViewDatas().put("task", new TaskForm(currentTask));
-
-        } catch (Exception e) {
-            viewModel.getErrors().add(e);
-        } finally {
-            viewModel.setTemplate("projects:details_project_tasks_config.html");
-
-            viewModel.getViewDatas().put("tasks", this.projectService.listProjectTasks(actor, project));
-            viewModel.getViewDatas().put("taskTypes", this.projectService.listTaskType());
-            viewModel.getViewDatas().put("project", project);
-            viewModel.getViewDatas().put("allTaskStatus", TaskStatus.values());
-            viewModel.getViewDatas().put("allProjectMilestones", this.projectService.listProjectMilestones(actor, project));
-
-            /* Get datas for line-chart*/
-            if (currentTask.getId() != null) {
-                this.getDatasForCharts(actor, viewModel, currentTask);
-            }
-        }
-    }
-
-    private Task createTask(User actor, Project project, HttpServletRequest request) throws ParseException, BusinessException {
-        TaskForm taskForm = new TaskForm(request);
-        Milestone milestone = taskForm.getMilestoneID() != null ? this.projectService.getMilestoneById(actor, taskForm.getMilestoneID()) : null;
-
-        return this.projectService.createTask(actor,
-                project,
-                taskForm.getTaskName(),
-                taskForm.getTaskComment(),
-                taskForm.getStartDate(),
-                taskForm.getEndDate(),
-                taskForm.getOriginalEstimate(),
-                taskForm.getTaskTypeID(),
-                this.userService.findUserByID(taskForm.getAssignedUserID()),
-                ProjectService.ORIGIN_TIMEBOARD,
-                null,
-                null,
-                milestone
-                );
-    }
-
-    private Task updateTask(User actor, Project project, Task currentTask, HttpServletRequest request) throws ParseException, BusinessException {
-        TaskForm taskForm = new TaskForm(request);
-        Milestone milestone = taskForm.getMilestoneID() != null ? this.projectService.getMilestoneById(actor, taskForm.getMilestoneID()) : null;
-        User assigned = this.userService.findUserByID(taskForm.getAssignedUserID());
-
-        final TaskType taskType = this.projectService.findTaskTypeByID(taskForm.getTaskTypeID());
-        currentTask.setTaskType(taskType);
-        currentTask.setName(taskForm.getTaskName());
-        currentTask.setComments(taskForm.getTaskComment());
-        currentTask.setStartDate(taskForm.getStartDate());
-        currentTask.setEndDate(taskForm.getEndDate());
-        currentTask.setOriginalEstimate(taskForm.getOriginalEstimate());
-        currentTask.setMilestone(milestone);
-        currentTask.setAssigned(assigned);
-
-        return this.projectService.updateTask(actor, currentTask);
-    }
-
-    public static class TaskForm {
-        private Long taskID;
-        private String taskName;
-        private String taskComment;
-        private Date startDate;
-        private Date endDate;
-        private Double originalEstimate;
-        private Long assignedUserID;
-        private Long taskTypeID;
-        private User assignedUser;
-
-        private TaskType taskType;
-        private TaskStatus taskStatus;
-        private Long milestoneID;
-
-        public TaskForm(Task task) {
-            taskID = task.getId();
-            taskType = task.getTaskType();
-            milestoneID = task.getMilestone() != null ? task.getMilestone().getId() : null;
-
-            originalEstimate = task.getOriginalEstimate();
-            startDate = task.getStartDate();
-            endDate = task.getEndDate();
-            assignedUser = task.getAssigned();
-            taskName = task.getName();
-            taskComment = task.getComments();
-            taskStatus = task.getTaskStatus();
-            taskStatus = TaskStatus.PENDING;
-
-
-        }
-
-        public TaskForm(HttpServletRequest request) throws ParseException {
-            if (!request.getParameter("taskID").isEmpty()) {
-                taskID = Long.parseLong(request.getParameter("taskID"));
-            }
-            taskName = request.getParameter("taskName");
-            taskComment = request.getParameter("taskComments");
-            startDate = new Date(DATE_FORMAT.parse(request.getParameter("taskStartDate")).getTime() + (2 * 60 * 60 * 1000) + 1);
-            endDate = new Date(DATE_FORMAT.parse(request.getParameter("taskEndDate")).getTime() + (2 * 60 * 60 * 1000) + 1);
-            originalEstimate = Double.parseDouble(request.getParameter("taskOriginalEstimate"));
-            taskStatus = request.getParameter("taskStatus") != null ? TaskStatus.valueOf(request.getParameter("taskStatus")) : TaskStatus.PENDING;
-
-            if (request.getParameter("taskMilestoneId") != null && !request.getParameter("taskMilestoneId").isEmpty()) {
-                milestoneID = Long.parseLong(request.getParameter("taskMilestoneId"));
-            }
-
-            if (request.getParameter("taskTypeID") != null &&  !request.getParameter("taskTypeID").isEmpty()) {
-                taskTypeID = Long.parseLong(request.getParameter("taskTypeID"));
-            }
-
-            if (request.getParameter("taskAssigned") != null && !request.getParameter("taskAssigned").isEmpty()) {
-                assignedUserID = Long.parseLong(request.getParameter("taskAssigned"));
-            }
-        }
-
-        public Long getTaskID() {
-            return taskID;
-        }
-
-        public void setTaskID(Long taskID) {
-            this.taskID = taskID;
-        }
-
-        public String getTaskName() {
-            return taskName;
-        }
-
-        public String getTaskComment() {
-            return taskComment;
-        }
-
-        public Date getStartDate() {
-            return startDate;
-        }
-
-        public Date getEndDate() {
-            return endDate;
-        }
-
-        public Double getOriginalEstimate() {
-            return originalEstimate;
-        }
-
-        public Long getAssignedUserID() {
-            return assignedUserID;
-        }
-
-        public void setAssignedUserID(Long assignedUserID) {
-            this.assignedUserID = assignedUserID;
-        }
-
-        public Long getTaskTypeID() {
-            return taskTypeID;
-        }
-
-        public void setTaskTypeID(Long taskTypeID) {
-            this.taskTypeID = taskTypeID;
-        }
-
-        public TaskType getTaskType() {
-            return taskType;
-        }
-
-        public void setTaskType(TaskType taskType) {
-            this.taskType = taskType;
-        }
-
-        public User getAssignedUser() {
-            return assignedUser;
-        }
-
-        public void setAssignedUser(User assignedUser) {
-            this.assignedUser = assignedUser;
-        }
-
-        public TaskStatus getTaskStatus() {
-            return taskStatus;
-        }
-
-        public void setTaskStatus(TaskStatus taskStatus) {
-            this.taskStatus = taskStatus;
-        }
-
-        public Long getMilestoneID() {
-            return milestoneID;
-        }
-
-        public void setMilestoneID(Long milestoneID) {
-            this.milestoneID = milestoneID;
-        }
-    }
-}
