--- projects/src/main/java/timeboard/projects/ProjectsTaskValidationServlet.java
@@ -1,78 +1,0 @@
-package timeboard.projects;
-
-/*-
- * #%L
- * projects
- * %%
- * Copyright (C) 2019 Timeboard
- * %%
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- * 
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- * 
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- * #L%
- */
-
-import java.io.IOException;
-import javax.servlet.Servlet;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
-import org.osgi.service.component.annotations.ServiceScope;
-import timeboard.core.api.ProjectService;
-import timeboard.core.api.exceptions.BusinessException;
-import timeboard.core.model.Project;
-import timeboard.core.model.User;
-import timeboard.core.ui.TimeboardServlet;
-import timeboard.core.ui.ViewModel;
-
-@Component(
-        service = Servlet.class,
-        scope = ServiceScope.PROTOTYPE,
-        property = {
-                "osgi.http.whiteboard.servlet.pattern=/projects/tasks/validation",
-                "osgi.http.whiteboard.context.select=(osgi.http.whiteboard.context.name=timeboard)"
-        }
-)
-public class ProjectsTaskValidationServlet extends TimeboardServlet {
-
-    @Reference
-    public ProjectService projectService;
-
-    @Override
-    protected ClassLoader getTemplateResolutionClassLoader() {
-        return ProjectsTaskValidationServlet.class.getClassLoader();
-    }
-
-
-    @Override
-    protected void handleGet(User actor, HttpServletRequest request, HttpServletResponse response, ViewModel viewModel) throws ServletException, IOException, BusinessException {
-        long id = Long.parseLong(request.getParameter("projectID"));
-        Project project = this.projectService.getProjectByID(actor, id);
-
-        viewModel.getViewDatas().put("tasks", this.projectService.listProjectTasks(actor, project));
-        viewModel.getViewDatas().put("project", project);
-        viewModel.setTemplate("projects:details_project_tasks_validation.html");
-        viewModel.getViewDatas().put("task", false);
-
-
-
-
-    }
-
-}
