--- projects/src/main/resources/templates/details_project_tasks_config.html
@@ -1,196 +1,0 @@
-<!DOCTYPE html>
-
-<html layout:decorate="~{projects:details_project_tasks.html}"
-      xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout"
-      xmlns:th="http://www.thymeleaf.org">
-
-<head>
-    <title>Project Tasks details</title>
-</head>
-<body>
-
-<div layout:fragment="task-config">
-
-    <form action="" class="ui form segment" enctype="application/x-www-form-urlencoded" id="task"
-          method="post" style="margin-left:1em;">
-        <input name="taskID" required th:value="${task.taskID}" type="hidden">
-        <div class=" two fields">
-            <div class="field">
-                <label>Task Name</label>
-                <input name="taskName" placeholder="Task Name" required th:value="${task.taskName}" type="text">
-            </div>
-            <div class="field">
-                <label>User assigned</label>
-                <div class="ui category search">
-                    <div class="ui icon input">
-
-                        <input class="prompt assigned" placeholder="Search user..." th:if="${task.assignedUser != null}"
-                               th:value="${task.assignedUser.name}" type="text"/>
-                        <input class="prompt assigned" placeholder="Search user..." th:if="${task.assignedUser == null}"
-                               type="text"/>
-
-                        <input class="taskAssigned" name="taskAssigned" th:if="${task.assignedUser != null}"
-                               th:value="${task.assignedUser.id}" type="hidden"/>
-                        <input class="taskAssigned" name="taskAssigned" th:if="${task.assignedUser == null}" type="hidden"
-                               value=""/>
-
-                        <i class="search icon"></i>
-                    </div>
-                    <div class="results"></div>
-                </div>
-            </div>
-        </div>
-
-        <div class="field">
-            <label>Task Description</label>
-            <textarea name="taskComments" placeholder="Task Description" th:text="${task.taskComment}" rows="2"></textarea>
-        </div>
-        <div class="two fields">
-            <div class="field">
-                <label>Task Start Date</label>
-                <input name="taskStartDate" placeholder="Task Name"
-                       required th:value="${#dates.format(task.startDate, 'yyyy-MM-dd')}" type="date">
-            </div>
-            <div class="field">
-                <label>Task End Date</label>
-                <input name="taskEndDate" placeholder="Task Name" required
-                       th:value="${#dates.format(task.endDate, 'yyyy-MM-dd')}" type="date">
-            </div>
-        </div>
-        <div class="four fields">
-        <div class="field">
-            <label>Task Original Estimate</label>
-            <input name="taskOriginalEstimate" placeholder="Task OE" required th:value="${task.originalEstimate}" type="number">
-        </div>
-        <div class="field">
-            <label>Task type</label>
-            <select  name="taskTypeID" >
-                <option th:each="type: ${taskTypes}"
-                        th:text="${type.typeName}"
-                        th:selected="${task.taskType != null && (task.taskType.id == type.id)}"
-                        th:value="${type.id}"></option>
-            </select>
-        </div>
-        <div class="field">
-            <label>Task status</label>
-            <select name="taskStatus" th:disabled="${task.taskID == null}">
-                <option th:each="status: ${allTaskStatus}"
-                        th:text="${status.label}"
-                        th:value="${status}"
-                        th:selected="${status == task.taskStatus}"></option>
-            </select>
-        </div>
-
-
-        <div class="field">
-            <label>Milestone assigned</label>
-            <select name="taskMilestoneId">
-                <option></option>
-                <option th:each="milestone: ${allProjectMilestones}"
-                        th:text="${milestone.name}"
-                        th:value="${milestone.id}"
-                        th:selected="${milestone.id == task.milestoneID}"></option>
-            </select>
-        </div>
-        </div>
-
-        <button class="ui positive button" type="submit">Save</button>
-
-        <a class="ui negative basic button" th:if="${task.taskID != null}"
-           th:href="@{/projects/tasks/delete(taskID=${task.taskID}, projectID=${project.id})}">
-            <i class="remove alternate icon"></i>
-            Delete Task
-        </a>
-
-    </form>
-
-    <div>
-<
-        <div class="ui fluid card" th:if="${task.taskID != null}">
-            <canvas id="lineChart" width="800" height="450"></canvas>
-        </div>>
-
-        <script th:inline="javascript">
-
-            /*<![CDATA[*/
-
-            const chart = new Chart(document.querySelector("#lineChart"), {
-              type: 'line',
-              data: {
-                labels: /*[[${listOfTaskDates}]]*/,
-                datasets: [{
-                    data: /*[[${effortSpentDatasForChart}]]*/,
-                    label: "Effort spent for " + /*[[${task.taskName}]]*/,
-                    borderColor: "#3e95cd",
-                    fill: true,
-                    steppedLine: true
-                  },
-                  {
-                    data: /*[[${realEffortDatasForChart}]]*/,
-                    label: "Real effort for " + /*[[${task.taskName}]]*/,
-                    borderColor: "#ff6384",
-                    fill: true,
-                    steppedLine: true
-                  }
-                ]
-              },
-              options: {
-                title: {
-                  display: true,
-                  text: 'Task - Real Effort and Effort Spent graph'
-                },
-                scales: {
-                    yAxes: [{
-                        ticks: {
-                            min: 0
-                        },
-                        scaleLabel: {
-                            display: true,
-                            labelString: 'Number of days'
-                        }
-                    }],
-                    xAxes: [{
-                        scaleLabel: {
-                            display: true,
-                            labelString: 'Dates'
-                        }
-                    }],
-                }
-              }
-            });
-
-            /*]]>*/
-
-
-            </script>
-    </div>
-
-    <script>
-    $('select.dropdown')
-      .dropdown()
-    ;
-
-         $('.ui.search')
-            .search({
-                apiSettings: {
-                        url: '/search?q={query}&projectID=[[${project.id}]]'
-                    },
-                    fields: {
-                        results : 'items',
-                        title   : 'screenName',
-                        description : 'email'
-                    },
-                    onSelect: function(result, response) {
-                        $('.assigned').val(result.screenName);
-                        $('.taskAssigned').val(result.id);
-                     },
-                minCharacters : 3
-            });
-
-
-    </script>
-
-</div>
-
-</body>
-</html>
